trickle: Retry encoder if necessary (#454)
This requires adding a last value cache for the media metadata
rather than pulling the metadata from a queue, because the queue
may be empty when the encoder restarts.

We also need to explicitly close the write end of any pipes that
may be dangling after the exception, since pyav will no longer be
closing those.