lv2v: Omit misleading healthcheck stacktrace (#476)
* lv2v: Omit misleading healthcheck stacktrace

It has been misleading investigations to include the stack trace
of the health check failure, since it looks like there is where
the problem is on the code. Since the exceptions we get here are
generally just problems with the healthcheck itself like the
server being down or hanging, we better just log the error message.

* lv2v: Add [HEALTHCHECK] prefix so it's easier to find in logs