main: Configure logging on the same level based on env (#477)
We were leaving the log_level to the default, which for
some reason meant we were only logging ERROR level onwards
(even tho python docs say it would log everything).

I didn't spend too much time understanding why that was the
case, but configuring the logs like this fixes the issue and
we get all INFO logs (or DEBUG if verbose is set). So we'll
be in a better place to debug the infer.py crashes.