trickle: Start subscribing at the current seq, rather than the next (#547)
This helps mitigate a bit of delay if the runner starts
subscribing to a channel after the first segment has already
started writing. This would have made the runner start waiting
for the *next* segment (1s on average, up to 2s), which seems
to happen roughly 2/3s of the time (eyeballing a histogram).