live/streamer: Improve monitor loop reliability (#568)
* live/streamer: stop_stream -> on_stream_stopped

* live/streamer: on_before_process_restart -> trigger_stop_stream

Had to refactor the exiting stop/wait methods a bit to avoid
redundancy.

* live/streamer: Move input_timeout logic to process_guardian

Also make it non-configurable. It was only necessary when we
ran infer.py for the old realtime playground where the streamer
was never stopped. Simplify this.

* live/streamer: on_stream_stopped -> callbacks.is_stream_running

* live/streamer: Create INITIALIZING state on process

* live/streamer: Fix pipeline init error handling

We were sometimes returning None in case we tried loading
with default params. Simple fix.

* live/streamer: Centralize state logic on _compute_current_state

Merged logic from monitor loop and get_state into the previously
named _current_state. Renamed to _compute_current_state.

Right now it's mostly mixing up all the logic between those sources,
needs refactoring so it actually works. Monitor loop is the one
responsible for updating the state periodically, and will restart
process in case we get an ERROR state

* live/streamer: Implement new compute_current_state logic

* live/streamer: Rename initial state to LOADING and always log changes

* live/streamer: Don't come back from ERROR automatically

* live/streamer: Fix monitor logic initial bugs

- The is_stream_running was async by accident which was causing an ERROR state on the start
- We were failing right in the beginning of the stream cause it hit the input timeout of 60s. Reordered monitor loop to make sense
- We were resetting the status object on reset_stream. We shouldn't do that, but only reset the InputState

* streamer: Improve logging on process restart logic

only enter on the first time the state changes to ERROR

* streamer: Move stream stop logic to monitor loop

Since we're not calling _restart rn we need it outside
to make sure we still stop the stream

* live/api: Add LOADING state support on healthcheck (#576)

* streamer: Fix potential race condition on monitor_loop

We would be temporarily setting the state to ERROR until we
actually restarted the process. To avoid ever returning ERROR
to the worker let's only set it if restart fails.

* live/streamer: Make sure we set ONLINE when starting streamer