live/streamer: Calculate FPS using a rolling window (#580)
* live/streamer: Calculate FPS using a rolling window

* live/streamer: Make sure FPS is kept up to date

Call it from the monitor loop

* live/streamer: Rename some variables

* live/streamer: Simplify FPS calculation

* live/streamer: Remove redundant FPS calculation on streamer

* Fix typing on streamer

* live/streamer: Create class to encapsulate FPS calculation

Fixes the problem of biased FPS measuremenets when frames
only start to come by the end of the window.

* live/streamer: Fix reset logic of FPSCounter

* Revert relative import change

* Keep FPS logs in the same format as before

We already have monitoring built around them so let's not change

* Only log FPS if stream is running

* Fix FPS calculation

We should not do -1 on frame count if we have a fixed measurement window.

* Only log FPS if it's >0